+++ app/src/main/java/com/example/daytask/activity/AuthActivity.kt
@@ -1,4 +1,4 @@
-package com.example.daytask
+package com.example.daytask.activity
 
 import android.content.Intent
 import android.os.Bundle
