+++ app/src/main/java/com/example/daytask/activity/SplashActivity.kt
@@ -1,5 +1,6 @@
-package com.example.daytask
+package com.example.daytask.activity
 
+import android.annotation.SuppressLint
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,40 +9,27 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
+import com.example.daytask.ui.screens.splash.SplashScreen
 import com.example.daytask.ui.theme.DayTaskTheme
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
 
-class MainActivity : ComponentActivity() {
-    private lateinit var auth: FirebaseAuth
+@SuppressLint("CustomSplashScreen")
+class SplashActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        auth = Firebase.auth
-        val user = intent.extras?.getString("user")
-
         setContent {
             DayTaskTheme {
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    DayTaskApp(
-                        email = user!!,
-                        signOut = {
-                            auth.signOut()
-                            goBackToAuth()
+                    SplashScreen(
+                        nextActivity = {
+                            startActivity(Intent(this, AuthActivity::class.java))
+                            finish()
                         }
                     )
                 }
             }
         }
-    }
-
-    private fun goBackToAuth() {
-        val intent = Intent(this, AuthActivity::class.java)
-        startActivity(intent)
-        finish()
     }
 }
